+++ src/vconsole.js
@@ -13,6 +13,9 @@ Unless required by applicable law or agreed to in writing, software distributed 
  * A Front-End Console Panel for Mobile Webpage
  */
 
+// global
+import './lib/symbol.js';
+
 // classes
 import VConsole from './core/core.js';
 import VConsolePlugin from './lib/plugin.js';