+++ CHANGELOG.md
@@ -6,6 +6,7 @@ English | [简体中文](./CHANGELOG_CN.md)
 - [FEATURE] Add `onReady` and `onClearLog` callback function to `vConsole.option`.
 - [FEATURE] Auto clear logs when `console.clear()` is called.
 - [FIX] Fix `\r` error when build in Windows.
+- [FIX] Fix `Symbol` error in iOS8 or other old OS.
 
 
 #### V3.0.0 (2017-09-27)